LPD-38548 feat implement batch contributor
